+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/ShaderPass/ShaderPass.cs
@@ -24,6 +24,10 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         DbufferProjector,
         DbufferMesh,
         ForwardEmissiveProjector,
-        ForwardEmissiveMesh
+        ForwardEmissiveMesh,
+        Raytracing,
+        RaytracingReflection,
+        RaytracingVisibility,
+        RaytracingForward,
     }
 }